Major Refactor
Refactored capture and input logic. Script no longer needs focus on Chiaki | Stream window. So it can just run in the background. As long as it's not minimized.
Added additional checks for race start screen and next race screen so it won't get stuck on those screens.
Refactored input keys. keys can now be configured in the config.py. These should match with what is set in chiaki!
Remove dependencies for the following.
- Removed the need for autohotkey.exe to be installed.
- Removed the need the following pip packages. (ahk, mss)

Added new pip dependency
- pywin32

pynput is currently only needed in screenshot.py not for the gran_turismo.py
updated readme for new changes